+++ ApiSGCOlimpiada/ApiSGCOlimpiada/Models/ProdutoPedidoOrcamento.cs
@@ -8,11 +8,11 @@ namespace ApiSGCOlimpiada.Models
     public class ProdutoPedidoOrcamento
     {
         public long Id { get; set; }
-        public double valor { get; set; }
-        public int Quantidade { get; set; }
-        public double Ipi { get; set; }
-        public double Icms { get; set; }
-        public double Desconto { get; set; }
+        public decimal valor { get; set; }
+        public decimal Quantidade { get; set; }
+        public decimal Ipi { get; set; }
+        public decimal Icms { get; set; }
+        public decimal Desconto { get; set; }
         public decimal TotalItem { get; set; }
         public Orcamento Orcamento { get; set; }
         public ProdutoSolicitacao ProdutoSolicitacao { get; set; }
